VXFM-6101 Support Switch firmware upgrade by compacting the firmware files in cases of low space

+++ lib/puppet/provider/cisconexus5k_firmwareupdate/cisco.rb
@@ -42,6 +42,22 @@ Puppet::Type.type(:cisconexus5k_firmwareupdate).provide :cisconexus5k, :parent =
     Puppet.debug("*** Cisco Version*********")
     Puppet.debug(installedversion)
 
+    Puppet.debug("*** Deleting unused Binaries, if any. *********")
+    delete_unused_binaries
+
+    compact_supported = supports_compact(installedversion)
+
+    if !space_available_to_copy(path) && compact_supported
+      Puppet.debug("*************************************************")
+      Puppet.debug("Compacting Installed Cisco Switch firmware binary")
+      Puppet.debug("*************************************************")
+      installed_image_file_name = parse_image_file(send_command("show version | grep image"))
+      send_command("install all nxos " + installed_image_file_name + " compact") do |out|
+        Puppet.debug("Image File Compact result: " + out)
+      end
+      Puppet.debug("*** Installed Image File Compacted ***")
+    end
+
     Puppet.debug("Flash file path ::" + flashfilepath)
     copysuccessful = copy_binary_to_switch(url,filename)
     unless copysuccessful
@@ -50,44 +66,52 @@ Puppet::Type.type(:cisconexus5k_firmwareupdate).provide :cisconexus5k, :parent =
       raise err
     end
 
+    Puppet.debug("Copy startup config")
     if(!copy_startup_config())
       err = "Unable to update firmware version as copy startup config failed"
       Puppet.debug("*********"+err+"*********")
       raise err
     end
-    if is_upgradable(flashfilepath)
-      Puppet.debug("**************************************")
-      Puppet.debug("Upgrading Cisco Switch firmware")
-      Puppet.debug("**************************************")
-      confirminstall=false
-      out = send_command("install all nxos "+ flashfilepath +"  non-interruptive")
-      responsetxt ="firmware update is successfull"
-    else
-      err = "Unable to update firmware as the version matches the current installed version"
-      Puppet.debug("*********"+err+"*********")
-      raise err
+
+    Puppet.debug("**************************************")
+    Puppet.debug("Upgrading Cisco Switch firmware")
+    Puppet.debug("**************************************")
+
+    installout = ''
+    command = "install all nxos "+ flashfilepath + " non-interruptive"
+    send_command(command) do |out|
+      installout << out
     end
 
+    unless installout.scan("Use compact image").empty?
+      Puppet.debug("System will compact the image before installing")
+      compact_binary flashfilepath
+      installout = ''
+      command = "install all nxos "+ flashfilepath + " non-interruptive"
+      send_command(command) do |out|
+        installout << out
+      end
+    end
+
+    unless installout.scan("Finishing the upgrade").empty?
+      responsetxt = "firmware update is successfull"
+      Puppet.debug("Waiting for switch to reboot")
+      sleep 180
+    end
     return responsetxt
   end
 
-  def is_upgradable(filepath)
-    successresponse=''
-    send_command("show install all impact nxos "+ filepath) do |out|
-      failresponse = out.scan('Pre-upgrade check failed')
-      unless failresponse.empty?
-        return false
-      end
-      successresponse << out
+  def parse_image_file(fileinfo)
+    installed_image_file_details = fileinfo.scan(/NXOS.+bin/i)[0]
+    Puppet.debug("Installed Image File Details = " + installed_image_file_details)
+    installed_image_file_details.split("\/").last
+  end
+
+  def compact_binary(filepath)
+    compactresponse = ''
+    send_command("install all nxos "+ filepath +" compact") do |out|
+      compactresponse << out
     end
-    versionimpact = successresponse.match(/^nxos:\s+(\S+)$|nxos \s+(.*?)$/m)
-    Puppet.debug("************Version Comparision************")
-    Puppet.debug(versionimpact)
-    Puppet.debug("*************************")
-    unless versionimpact[0].scan("yes").empty?
-      return true
-    end
-    return false
   end
 
   def copy_binary_to_switch(url,filename)
@@ -98,8 +122,6 @@ Puppet::Type.type(:cisconexus5k_firmwareupdate).provide :cisconexus5k, :parent =
       readdir<<out
     end
     filenamefound = readdir.scan(filename)
-    Puppet.debug("**********filenamefound**********")
-
     if filenamefound[0]==filename
       Puppet.debug("File already exists, checking image version")
       deleteout = ''
@@ -119,25 +141,25 @@ Puppet::Type.type(:cisconexus5k_firmwareupdate).provide :cisconexus5k, :parent =
     send_command("dir bootflash:", :prompt => /#{filename}/) do |out|
       existreaddir<<out
     end
-      filefound = existreaddir.scan(filename)
-      Puppet.debug("******  File found ******")
-      Puppet.debug(filefound)
+    filefound = existreaddir.scan(filename)
+    Puppet.debug("******  File found ******")
+    Puppet.debug(filefound)
 
-   if filefound[0]==filename
+    if filefound[0]==filename
       copysuccess=true
       Puppet.debug("Successfully copied the file")
     end
-   return copysuccess
+    return copysuccess
   end
 
   def copy_startup_config()
     updateout = ''
-	copystartupconfigresponse = ''
+    copystartupconfigresponse = ''
     send_command("copy running-config startup-config") do |out|
       updateout<<out
     end
-	Puppet.debug("updateout copy startup config")
-	Puppet.debug(updateout)
+    Puppet.debug("updateout copy startup config")
+    Puppet.debug(updateout)
     return true
   end
 
@@ -154,12 +176,65 @@ Puppet::Type.type(:cisconexus5k_firmwareupdate).provide :cisconexus5k, :parent =
     if !File.exist? http_dir
       FileUtils.mkdir_p http_dir
     end
-    FileUtils.cp path, full_http_path
+    if !File.exist? full_http_path
+      FileUtils.cp path, full_http_path
+    end
     FileUtils.chmod_R 0755, http_dir
   end
 
+  def delete_unused_binaries
+    out =  send_command("dir bootflash: | grep nxos")
+    bin_file_arr =  out.scan(/nxos\S+bin/)
+    bin_file_arr.each do |bin_file|
+      send_command("delete bootflash:"+bin_file+" no-prompt")
+    end
+  end
+
+  # File compact is attempted only when the free space is less than the size of the
+  # image file to be copied. Compact is supported only on version nxos.7.0.3.I3.1.bin
+  # or greater than nxos.7.0.3.I3.1.bin and on 31xx series switches.
+  def supports_compact(installedversion)
+    Puppet.debug("*** In supports_compact?(). installedversion - %s " %[installedversion])
+    compact_supported = false
+    out = send_command("show version | grep Chassis")
+    cisco_switch_details = ""
+    return false unless out =~ /cisco.+Chassis/i
+
+    cisco_switch_details << out.scan(/cisco.+Chassis/i).first
+    Puppet.debug(cisco_switch_details)
+    cisco_switch_series = cisco_switch_details.split(" ")[2]
+    Puppet.debug(cisco_switch_series)
+    if (cisco_switch_series =~ /31\d+./i)
+      min_version_supporting_compact = "7.0(3)I3(1)"
+      installed_version_num = installedversion.gsub(/[^\d]/, '')
+      Puppet.debug("Installed version number= %s " % [installed_version_num])
+      min_version_supporting_compact_num = min_version_supporting_compact.gsub(/[^\d]/, '')
+      if(installed_version_num.to_i >= min_version_supporting_compact_num.to_i)
+        Puppet.debug("Installed version supports Compacting...")
+        compact_supported = true
+      end
+    end
+    compact_supported
+  end
+
+  def space_available_to_copy(path)
+    space_available = false
+    Puppet.debug("*** In space_available_to_copy?(). File - %s " %[path])
+    file_size = File.size(path)
+    Puppet.debug(file_size)
+    free_space_details = ""
+    out = send_command("dir bootflash: | grep free")
+    free_space_details << out.scan(/\d+\s+bytes free/i).first
+    free_space = free_space_details.split(" ").first
+    Puppet.debug("Space available on the switch= %s " %[free_space])
+    if(free_space.to_i > file_size)
+      space_available = true
+    end
+    space_available
+  end
 
   def flush
     Puppet.debug("in firmwareupdate flush method")
+    super
   end
 end
